Othello is working on a single computer, need to add server compatibility
Had to do a lot of googling to see what I was doing wrong but, after rewriting it and abandoning the board api as not applicable, Othello is back and looking good thanks to the documentation on using image files in place of coloring JButtons. Need a bit of help making it talk to the server in the way we want it to but as far as the game functioning properly and being attractive, I think we're good. Features a score counter, win checking, a thing which shows valid moves (because it's way easier to tell someone where they're allowed to click instead of waiting for a click and then checking), and a restart button which works at any point in the game.